// src/app/models/user/user.ts
export class User {
    id: number;
    userName: string;
    password: string;
    name: string;
    surName: string;
    emailAddress: string;
    statusId: number;
    personId: number;
    image: string;
    creationTime: string;
    creatorUserId: string;
    lastModificationTime: string;
    lastModifierUserId: number;
    deleterUserId: number;
    deletionTime: string;
    isActive: boolean;
    isDeleted: boolean;
}